requirements.txt
mysql-connector==2.2.9
mysql-connector-python==8.0.31
protobuf==3.20.1
